First Strike!
I set up the random foil collision location. It was easy enough. The one modification from the original code I made was to drop the assumption that the foil subtends a small angle and let the central locations be slightly more likely to be struck, just because accounting for that isn't that hard.