I want calculus! That is my simple request.
All things calculus, bigger and better and best!

I did the part that integrates the stopping power to make the ion slow down in the foil. It took a while, but I think the solution up with which I came is efficient and elegant. I will want to confirm I got all the math right, but for now, I'm ready to move on to the polynomials!