Discerning
I realized that a lot (though definitely not all) of my error was coming from extremely hot or extremely cold deuterons. The COSY mapping (and the inverse mapping I based on it) only work on deuterons near 12 MeV, and the CsI cathode is only rated for energies from 10.7 to 14.2 MeV. Therefore, I've stripped all particles outside of those bounds as irrelevant and also poorly simulated. This upped the speed of the progam by a good 5 decibels, and also reduces a lot of apparent distortion in the inferred spectrum.