One order to go
I did time correction, and it works to second order or so. Problem is, I need accuracy to third order or so to get 20 ps accuracy. It looks like there's some correctable error in the energy determination, and I think if I get rid of that, aberrations will bring the total error to right around 20 ps.