>exasperated sigh<
So the deconvolution was nowhere near functional. So I turned to fitting. I'm using some Nelder-Mead stuff because I don't want to deal with gradients. It's not very good. It is, in fact, hot garbage. Why is this happening? I don't understand why it's such garbage. I was so careful about the spectrum model and the optimization scheme. aaaaaa.