Partial garbage!
I've finally suceeded in getting a fit that's not complete garbage! I drastically reduced the number of degrees of freedom on each of the profiles so that it couldn't really overfit anymore. There were some serious nonlinearities that forced be to be clever at avoiding degeneracies in those parameters, though. The fits are still pretty basic and, interestingly enough, not using all of the degrees of freedom I've given it. There's lots of room for improvement, but let's just take a second to appreciate the fact that my fits look like real fits for once.