Aaaaaa
Apache was garbage so I decided to find an L-BFGS-B solver online but the only one I could find wasn't really working either so I decided to just code it myself and I've been bugfixing this for days. I'm tempted to blame myself, but the original paper by Nocedal et. al. is actually riddled with typos. Bad typos. Like, sign errors. WTF‽ Anyway it sort of sometimes works now? God, it's so inefficient because of all the redundancies and assertions. It's not any good.