Nocedal is a dirty liar
I think I've gotten it to the point where it consistently finishes minimizing without any errors. There were so many roundoff issues! I don't understand how Nocedal et al was ever able to run code like this! If only their code was readable. Anyway, this way is less efficient but has a lot less of that. I only pray that it's little enough. The results are still trash so probably not but who knows. I'm going to start clearing out all the print statements and identify the step that's limiting the speed now.