I should stop touching it now
I think I've gotten about as good as I can get with this for now. I've recently learned of a completely different algorithm that may yield similar results with much less iterations and fiddling, but that is a later issue. For now, I need to figure out what's going on with the mysterious double temperature thing.