Ehhh?
I've managed to get rid of most of the most glaring artifacts of ultra low yield, but there's only so much I can do. At the end of the day, a x1e-3 spectrum just doesn't have a lot of information. I think short of accounting directly for the upscatter spectrum, there's just not a lot I can do.