#ResearchFlatRhoR
The rho R is a lot flatter than I expected! I managed to finally get a pretty good curve by replacing my smoothening term with an entropy term. Entropy. Is there anything it can't do? I still want to play with it a bit. I'm not sure if all this nonlinear dependence is entirely healthy for it.