One less excuse
I made it so changing the asymmetry parameter does not change the down scatter yield in the optimization, in the hopes that it would help the optimization algorithm actually touch that term. It didn't work. I have confirmed that there is a not insignificant gradient in the modified asymmetry term now, though. It's just a matter of finding out why the algorithm isn't _touching_ it.