Well this is awful
I don't understand why the temperature insists on being like this even with no priors in place. I thought I had it for a second there, too, but it turns out that was statistics playing cruel tricks on me (as it turns out, this _isn't_ pseudorandom, and I don't know why). I wonder if I should changethe spectrum model to a true gaussian in energy space.