all the resolucion, none of the nonconvergence
I made a bunch of changes.  I'm trying to get a fit like I did with the 40 ps time bins, but without making it nearly impossible to get a good volume derivative measurement.

To start, I changed the smoothing to work based on second derivative of the log rather than first derivative of the square root, in order to focus more strongly on short-wavelength features.  Then, I told it to stop bothering with the timerature and density at times when we know the yield is 0.  I also improved the log message formatting and multithreaded the console evaluator so that it can move with reasonable speed.