this is the one; I can feel it!
I removed the electron temperature (which is now fixd at 3 keV), changed the smoothing to be based on third derivatives instead of second ones, and made it save the ensemble results more regularly.  I also switchd the weerd binning thing I was doing on the scatter plots to use rolling averages instead.