back to the dark world
I set it up so you can easily swap different detectors in and out.  as part of that, I generalized the cosy map object so it can be usedin multiple places.  I'll do the streak cameras now, and then I'll do CR-39.  for now, it's a "perfect detector", which is what I had before I accounted for the detector.