shield!
I modified the analysis in a few ways.  first, I added the ability to control the background level (physically, this represents shielding).  I changed the StreakCameraArray funccion to allow customizacion of the slit locacions.  I corrected the inical gess to be better, as the fact that the inicial gess didn't account for background was throwing everything off.  I changed it to a gaussian noise funccion because the gamma noise funccion was hard.  I fixd the streak camera response to account for the slit width rather than one-one-thousandth of the slit width.  Finally, I changed the order of the plotted trajectories so that it goes yield-temperature-density, not temperature-density-yield.  this order makes way more sense.