just ignore them
I actually mostly fixd the error bars!  it turns out that leaving the components I didn't optimize in the covariance calculacion just kind of broke it, I gess.  the error bars on the edge are still obviously rong, but I'm leaving it as it is for now because this is actually passable.