bowties are kakkowarui
I added a way to plot the focal plane automatically.

I also changed the thread implementation to use a nice bilt-in executor thingy, and also fixd a bunch of bugs that were causing it to fail whenever it was calld many times, which it was due to my new implementation.
 consistency.