New focal plane just dropped
jk not really.  it's supposed to be the same as it was before.  I just didn't back up my exact previous numbers, and it really all aut to be consistent.  oh, but I did move one streak camera lower in energy, since it seemd like it was picking up a lot of primary signal.