Binomi-él? I hardly know él!
I fixd my binomial distribution, which was very rong (I forgot the binomal term).  I also touchd up the new plot I added because it was just broken.  it's nice now that it actually appears; I've identified three major issues that I aut to address today: the slit is in a not great place (fixd that already), the counting statistics from the signal electrons are not taken into account, and I'm not sure if the background counting statistics are rite either (I need to plot the background).  oh, I also need to run this at hier temperatures at some point.