under the hood
I added a new plot that looks at the actual parabolas that are being fit.  wait, sweet jesus is that really how downshifted the spectrum is by the foil?  jeezum.  whatever, the 2-slit config with huge slits and background turnd way down seems to be sufficient for a HTPD poster, I hope.