Aaand done
some last touch-ups for my paper.  I set the smoothing back to 1e-10 (because 1e-9 was actually starting to look really bad in my view; I hope I didn't need it), made the background more generick (because it was really not exact in the first place, anyway), and made it calculate dTi/dt at peak ρR instead of dρR/dt at peak ρR (not sure why that was like that in the first place, onestly).  also made the output in view_ensemble more readable and made it more resistant to changing numbers of rows.