a questionable tactic
Johan wanted me to look at adding constraints to the temperature based on the yield, so I did that.  and it kind of works.  but it also tries to match the yield to the temperature, and as a result gets the burn history shape rong.  and I really don't think this is a reasonable plan in the first place because it defeats the point of a dTi/dt diagnostic.  so I'm going to revert this now.