no more long array of strings
this time, I'm putting background on the PDDT.  I also laid the groundwork for turning to protons instead of deuterons, tho I'm afraid to test that.  oh, I also added units to quantities and completely reworked the way it outputs values to make it more scalable.  also, it messes with the foil thickness now.