the reality of the situation
MRS foils are circular.  I really don't know what possesed me to make the foil square.  it seems that this change combined with the moving the foil farther from the implosion, also in accordance to what is actually achievable, has made the efficiency much worse, but it also made the resolution a bit better.